turn extendr engine into separate crate. fixes #46.
